+++ plugin_tooling/src-tests/com/github/rustdt/tooling/ops/RacerOutputParserTest.java
@@ -11,19 +11,23 @@
 package com.github.rustdt.tooling.ops;
 
 import static melnorme.lang.tooling.CompletionProposalKind.Function;
+import static melnorme.lang.tooling.CompletionProposalKind.Let;
+import static melnorme.lang.tooling.CompletionProposalKind.Struct;
+import static melnorme.lang.tooling.CompletionProposalKind.Trait;
 import static melnorme.utilbox.core.Assert.AssertNamespace.assertFail;
 import static melnorme.utilbox.core.CoreUtil.listFrom;
 
 import java.util.List;
 
 import melnorme.lang.tests.CommonToolingTest;
-import melnorme.lang.tooling.CompletionProposalKind;
 import melnorme.lang.tooling.ToolCompletionProposal;
 import melnorme.lang.tooling.ToolingMessages;
+import melnorme.lang.tooling.ast.SourceRange;
 import melnorme.lang.tooling.completion.LangCompletionResult;
 import melnorme.lang.tooling.ops.FindDefinitionResult;
 import melnorme.lang.tooling.ops.OperationSoftFailure;
 import melnorme.lang.tooling.ops.SourceLineColumnRange;
+import melnorme.utilbox.collections.ArrayList2;
 import melnorme.utilbox.core.CommonException;
 
 import org.junit.Test;
@@ -43,6 +47,10 @@ public class RacerOutputParserTest extends CommonToolingTest {
 		};
 	}
 	
+	protected SourceRange sr(int offset, int length) {
+		return new SourceRange(offset, length);
+	}
+	
 	@Test
 	public void test() throws Exception { test$(); }
 	public void test$() throws Exception {
@@ -58,8 +66,8 @@ public class RacerOutputParserTest extends CommonToolingTest {
 			"MATCH BufRead;BufRead;519;10;/RustProject/src/xpto2.rs;Trait;pub trait BufRead : Read {\n"
 			, 
 			listFrom(
-				new ToolCompletionProposal(offset, 0, "BufReader", CompletionProposalKind.Struct, "xpto.rs"),
-				new ToolCompletionProposal(offset, 0, "BufRead", CompletionProposalKind.Trait, "xpto2.rs")
+				new ToolCompletionProposal(offset, 0, "BufReader", "BufReader", Struct, "xpto.rs"),
+				new ToolCompletionProposal(offset, 0, "BufRead", "BufRead", Trait, "xpto2.rs")
 			)
 		);
 		
@@ -68,7 +76,7 @@ public class RacerOutputParserTest extends CommonToolingTest {
 			"MATCH BufReader;BufReader;32;11;relativeDir/src/xpto.rs;Let;pub struct BufReader<R> {\n"
 			, 
 			listFrom(
-				new ToolCompletionProposal(offset-2, 2, "BufReader", CompletionProposalKind.Let, "xpto.rs")
+				new ToolCompletionProposal(offset-2, 2, "BufReader", "BufReader", Let, "xpto.rs")
 			)
 		);
 		
@@ -81,21 +89,26 @@ public class RacerOutputParserTest extends CommonToolingTest {
 					+"pub fn copy<R: Read, W: Write>(r: &mut R, w: &mut W) -> io::Result<u64> {\n"
 			, 
 			listFrom(
-				new ToolCompletionProposal(offset, 0, "stdin", "stdin()", Function, "stdio.rs"),
-				new ToolCompletionProposal(offset, 0, "repeat", "repeat(byte: u8)", Function, "util.rs"),
-				new ToolCompletionProposal(offset, 0, "copy", "copy(r: &mut R, w: &mut W)", Function, "util.rs")
+				new ToolCompletionProposal(offset, 0, "stdin", "stdin()", Function, "stdio.rs",
+					"stdin()", new ArrayList2<SourceRange>()),
+				new ToolCompletionProposal(offset, 0, "repeat", "repeat(byte: u8)", Function, "util.rs",
+					"repeat(byte)", new ArrayList2<>(sr(7, 4))),
+				new ToolCompletionProposal(offset, 0, "copy", "copy(r: &mut R, w: &mut W)", Function, "util.rs",
+					"copy(r, w)", new ArrayList2<>(sr(5, 1), sr(8, 1)))
 			)
 		);
 		
 		// Test error case
 		testParseOutput(buildParser, 
 			"PREFIX 1,1,\n" +
-			"MATCH stdin;stdin(;122;7;/rustc-nightly/src/libstd/io/stdio.rs;Function;pub fn stdin() -> Stdin {\n"+
-			"MATCH stdin2;stdin2(${;122;7;/rustc-nightly/src/libstd/io/stdio.rs;Function;pub fn stdin() -> Stdin {\n"
+			"MATCH xxx;xxx(;122;7;/rustc-nightly/src/libstd/io/stdio.rs;Function;pub fn stdin() -> Stdin {\n"+
+			"MATCH xxx2;xxx2(${;122;7;/rustc-nightly/src/libstd/io/stdio.rs;Function;pub fn stdin() -> Stdin {\n"
 			, 
 			listFrom(
-				new ToolCompletionProposal(offset, 0, "stdin", "stdin()", Function, "stdio.rs"),
-				new ToolCompletionProposal(offset, 0, "stdin2", "stdin2()", Function, "stdio.rs")
+				new ToolCompletionProposal(offset, 0, "xxx", "xxx()", Function, "stdio.rs",
+					"xxx()", new ArrayList2<SourceRange>()),
+				new ToolCompletionProposal(offset, 0, "xxx2", "xxx2()", Function, "stdio.rs",
+					"xxx2(__)", new ArrayList2<SourceRange>(sr(5, 2)))
 			)
 		);
 	}