+++ documentation/ChangeLog.md
@@ -1,10 +1,11 @@
 ## RustDT release ChangeLog
 
 ### next
- * Added: Code Completion functionality using Racer. TODO: Features pics
  * Added: Open Definition functionality using Racer.
-  * [Doc] Added section https://github.com/RustDT/RustDT/blob/latest/documentation/UserGuide.md#content-assistopen-definition
- * [Doc] Added section https://github.com/RustDT/RustDT/blob/latest/documentation/UserGuide.md#editor-newline-auto-indentation
+ * Added: Code Completion functionality using Racer. TODO: Features pics
+  * Added auto-insert of function arguments for function proposals.
+  * [Doc] Added section https://github.com/RustDT/RustDT/blob/latest/documentation/UserGuide.md#content-assistopen-definition .
+ * [Doc] Added section https://github.com/RustDT/RustDT/blob/latest/documentation/UserGuide.md#editor-newline-auto-indentation .
  * Added: Content Assist of code snippets. Configurable in `Preferences/Rust/Editor/Code Snippets`. TODO more defaults
 
 ### 0.1.1